separate the classes
