+++ book.js
@@ -1,0 +1,7 @@
+// new book constructor
+export default class Book {
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+  }
+}
