+++ index.html
@@ -30,8 +30,9 @@
                 <button id="submit" type="submit">Submit</button>
             </div>
         </form>
-        <script src="script.js">
-        </script>
+        <script type="module" src="script.js"></script>
+        
+        
 </body>
 
 </html>
