+++ style.css
@@ -1,89 +1,1 @@
 @import url('https://fonts.googleapis.com/css2?family=Cabin:ital,wght@0,400;0,500;0,600;0,700;1,400;1,500;1,600;1,700&family=Hind+Madurai:wght@300;400;500;600;700&display=swap');
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-body{
-  font-family: 'Cabin', sans-serif;
-}
-.container {
-  display: flex;
-  flex-direction: column;
-  margin-top: 40px;
-  margin-left: 20%;
-  margin-right: 20%;
-}
-
-h1 {
-  align-self: center;
-  margin-bottom: 15px;
-}
-
-hr {
-  border: 1px solid black;
-  width: 20%;
-  align-self: center;
-  margin-bottom: 20px;
-}
-
-.table,
-th {
-  border: 1.5px solid rgb(63, 41, 41);
-  border-collapse: collapse;
-  box-shadow: 1px 2px rgb(68, 1, 1);
-}
-
-.table {
-  margin-bottom: 15px;
-}
-
-th {
-  padding: 8px;
-}
-
-tr:nth-child(even) {
-  background-color: rgb(247, 179, 179);
-}
-
-.remove-btn {
-  width: 100%;
-  height: 30px;
-  justify-self: flex-end;
-  border-color: rgb(46, 0, 0);
-  box-shadow: 2px 4px rgb(27, 0, 0);
-  margin-right: 0.625rem;
-}
-
-.remove-btn:hover {
-  cursor: pointer;
-}
-
-#enter {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  margin-left: 20%;
-  margin-right: 20%;
-  margin-bottom: 50px;
-}
-
-#title,
-#author {
-  height: 25px;
-  padding: 5px;
-  border: 1px solid rgb(97, 3, 3);
-  box-shadow: 1px 2px rgb(58, 55, 55);
-}
-
-#submit {
-  width: 6rem;
-  border-color: rgb(87, 2, 2);
-  box-shadow: 2px 4px rgb(36, 0, 0);
-  align-self: flex-end;
-  height: 25px;
-}
-
-#submit:hover {
-  cursor: pointer;
-}
